xarm-gym-env-pybullet/requirements.txt
env==0.1.0
gym==0.26.2
numpy==1.25.2
pybullet==3.2.5
PyYAML==6.0.1
